Extract harness wallet code into own module
